requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-ckeditor==6.7.0
django-modeltranslation==0.18.11
redis==5.0.1
google-cloud-translate==3.12.0
pytest==7.4.3
pytest-django==4.7.0
flake8==6.1.0
gunicorn==20.1.0

